requirements.txt
# AI Career Guidance System - Dependencies

# Core ML/AI Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning (optional, for advanced models)
# torch>=2.0.0

# Natural Language Processing
nltk>=3.8.0
textblob>=0.17.0

# Microsoft Agent Framework for AI Chatbot
agent-framework-azure-ai

# OpenAI SDK
openai>=1.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Data Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Job Market Data (web scraping)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration & Environment
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# JSON Schema
jsonschema>=4.19.0

# Date/Time utilities
python-dateutil>=2.8.0

# Progress bars
tqdm>=4.66.0

# Testing
pytest>=7.4.0
